# api/openapi.yaml
openapi: 3.0.3
info:
  title: Secure Vault API
  description: |
    Zero-knowledge password manager API. The server stores and returns **only
    ciphertext + metadata** — it never sees plaintext vault items or master
    passwords. All encryption happens client-side (WASM crypto core, Phase 4).

    **API versioning:** v0 endpoints live under `/api`. Breaking changes in later
    phases will use `/api/v1`. Document any breaking change before moving the prefix.

    **Phase 3:** real session authentication replaces the Phase 2 `X-Dev-User` dev
    guard. Register derives the master key and auth hash client-side; the server
    stores only a server-side Argon2id hash of the auth hash (zero-knowledge
    invariant). Login issues a `vault_session` HttpOnly/Secure/SameSite=Strict cookie.
  version: 0.3.0
servers:
  - url: /
    description: Local development (served by cmd/server)
paths:
  /api/register:
    post:
      operationId: postRegister
      summary: Register a new account
      description: |
        Accepts the email, KDF params, client-derived auth hash, and the
        protected (wrapped) vault key. The server stores a server-side Argon2id
        hash of the auth hash — it never sees the master password or key.

        Returns a uniform 201 regardless of whether the email is already
        registered to prevent account enumeration.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created (or email already exists — response is identical).
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/login:
    post:
      operationId: postLogin
      summary: Authenticate and open a session
      description: |
        Accepts email + the client-derived auth hash. The server re-hashes the
        auth hash server-side and compares with the stored value using
        constant-time equality. On success, sets a `vault_session` session cookie
        (HttpOnly, Secure, SameSite=Strict) and returns the protected symmetric
        key + KDF params so the client can unlock the vault locally.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated. Session cookie set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/logout:
    post:
      operationId: postLogout
      summary: Revoke the current session
      description: |
        Revokes the session server-side and clears the `vault_session` cookie.
        Requires a valid session cookie and the `X-Vault-CSRF: 1` header.
      tags: [auth]
      responses:
        "204":
          description: Session revoked. Cookie cleared.
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/user:
    delete:
      operationId: deleteUser
      summary: Delete the authenticated user's account
      description: |
        Revokes all sessions, then permanently deletes the user row and all
        vault items (ON DELETE CASCADE). This is irreversible.
      tags: [auth]
      responses:
        "204":
          description: Account deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/user/key:
    put:
      operationId: putUserKey
      summary: Rotate master password (update auth credentials and wrapped key)
      description: |
        Atomically updates the server-side auth hash, KDF params, and protected
        symmetric key for a master-password change. The client re-derives
        everything locally and sends only the new auth hash + new wrapped vault
        key — the plaintext vault key and master key are never sent.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKeyRequest"
      responses:
        "200":
          description: Auth credentials and wrapped key updated.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/items:
    get:
      operationId: listItems
      summary: List the caller's vault items (ciphertext only)
      description: |
        Returns every vault item belonging to the authenticated user as opaque
        ciphertext + metadata. Requires a valid `vault_session` cookie (Phase 3).
      tags: [items]
      responses:
        "200":
          description: The user's encrypted vault items.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createItem
      summary: Store a new encrypted vault item
      description: |
        Accepts an opaque ciphertext blob + metadata only. The server validates
        the envelope shape but cannot read the contents.
      tags: [items]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewItem"
      responses:
        "201":
          description: Item stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/items/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: The item's server-assigned UUID.
    put:
      operationId: updateItem
      summary: Replace an existing encrypted vault item
      description: |
        Replaces the ciphertext and/or item_type of an existing item. Uses
        optimistic concurrency: the client supplies `current_revision` (the last
        revision it observed). The server rejects the write with 409 if the stored
        revision is newer, and increments `revision` + `updated_at` on success.
        This is last-write-wins guarded by revision — enough for two-device sync.
      tags: [items]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItem"
      responses:
        "200":
          description: Item updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Conflict — a newer revision exists. Fetch and merge before retrying.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteItem
      summary: Delete a vault item
      description: |
        Permanently deletes the item. The next `GET /api/sync` snapshot will not
        include it (full-snapshot model — no tombstones in v0).
      tags: [items]
      responses:
        "204":
          description: Item deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/sync:
    get:
      operationId: getSync
      summary: Full vault snapshot for unlock
      description: |
        Returns everything the client needs to unlock the vault: the user's
        protected symmetric key, the KDF params used to derive the master key,
        and the complete list of encrypted items.

        **Sync model (v0 — full snapshot):** returns ALL items on every call. No
        incremental sync, no pagination, no tombstones. A deleted item simply
        disappears from the next snapshot. Revisit only if a vault grows large
        enough to matter.
      tags: [sync]
      responses:
        "200":
          description: Full vault snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  responses:
    BadRequest:
      description: Invalid request body or parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found (or not owned by caller).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Rate limit exceeded. Back off and retry later.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Item:
      type: object
      required: [id, item_type, ciphertext, revision, updated_at]
      properties:
        id:
          type: string
          format: uuid
          description: Server-assigned item identifier.
        item_type:
          type: string
          description: Non-secret category hint (e.g. "login", "note", "card").
          enum: [login, note, card, identity]
        ciphertext:
          type: string
          description: |
            Base64-encoded (standard, padded) AEAD ciphertext of the item.
            Opaque to the server. Format: nonce(24 bytes) || ciphertext || tag(16 bytes),
            base64-encoded as a single blob. See plan/phase-1-crypto-core.md §wire-formats.
        revision:
          type: integer
          description: |
            Monotonically increasing server-side revision counter. Starts at 1 on
            creation, incremented by the server on every successful PUT. Used by
            clients for optimistic concurrency (supply in UpdateItem.current_revision).
          minimum: 1
        updated_at:
          type: string
          format: date-time
          description: Last revision timestamp (server clock).
    NewItem:
      type: object
      required: [item_type, ciphertext]
      properties:
        item_type:
          type: string
          enum: [login, note, card, identity]
        ciphertext:
          type: string
          description: Base64-encoded AEAD ciphertext. The server never decrypts it.
    UpdateItem:
      type: object
      required: [ciphertext, current_revision]
      properties:
        item_type:
          type: string
          enum: [login, note, card, identity]
          description: Optional — omit to keep existing item_type.
        ciphertext:
          type: string
          description: Replacement ciphertext blob (base64, standard padded).
        current_revision:
          type: integer
          description: |
            The revision the client last observed. The server rejects with 409 if
            the stored revision is strictly greater than this value.
          minimum: 1

    # KDFParams is the Argon2id derivation settings stored alongside each user.
    # The client receives these at sync time so it can re-derive the master key.
    # Wire format matches plan/phase-1-crypto-core.md §wire-formats exactly:
    # { "type": "argon2id", "version": 19, "memory_kib": 65536, "iterations": 3, "parallelism": 1 }
    KDFParams:
      type: object
      required: [type, version, memory_kib, iterations, parallelism]
      properties:
        type:
          type: string
          example: argon2id
        version:
          type: integer
          description: Argon2 version constant (0x13 = 19).
          example: 19
        memory_kib:
          type: integer
          description: Memory cost in kibibytes.
          example: 65536
        iterations:
          type: integer
          description: Time cost (iteration count).
          example: 3
        parallelism:
          type: integer
          description: Degree of parallelism.
          example: 1
    SyncResponse:
      type: object
      required: [protected_symmetric_key, kdf_params, items]
      properties:
        protected_symmetric_key:
          type: string
          description: |
            Base64-encoded wrapped vault key (AEAD-encrypted with the stretched
            master key). The client unwraps it after key derivation to obtain the
            vault key used to decrypt items.
        kdf_params:
          $ref: "#/components/schemas/KDFParams"
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
          description: Complete snapshot of all vault items (full-snapshot model, v0).
    RegisterRequest:
      type: object
      required: [email, kdf_params, auth_hash, protected_symmetric_key]
      properties:
        email:
          type: string
          format: email
          description: User's email address (normalized to lowercase before storage).
        kdf_params:
          $ref: "#/components/schemas/KDFParams"
        auth_hash:
          type: string
          description: |
            Base64-encoded (standard, padded) client-derived auth hash.
            This is the OUTPUT of the second independent Argon2id pass
            (DeriveAuthHash) — never the master password or master key.
        protected_symmetric_key:
          type: string
          description: |
            Base64-encoded AEAD-encrypted vault key (wrapped with the
            stretched master key). Stored opaquely; returned at sync time.
    LoginRequest:
      type: object
      required: [email, auth_hash]
      properties:
        email:
          type: string
          format: email
        auth_hash:
          type: string
          description: Base64-encoded client-derived auth hash (same derivation as registration).
    LoginResponse:
      type: object
      required: [protected_symmetric_key, kdf_params]
      properties:
        protected_symmetric_key:
          type: string
          description: |
            Base64-encoded wrapped vault key. The client unwraps it with the
            stretched master key to obtain the vault key for item decryption.
        kdf_params:
          $ref: "#/components/schemas/KDFParams"
    UpdateKeyRequest:
      type: object
      required: [auth_hash, kdf_params, protected_symmetric_key]
      properties:
        auth_hash:
          type: string
          description: New base64-encoded client-derived auth hash (re-derived from new password).
        kdf_params:
          $ref: "#/components/schemas/KDFParams"
        protected_symmetric_key:
          type: string
          description: New base64-encoded wrapped vault key (re-wrapped with new master key).
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Generic, non-sensitive error message safe to show clients.